// demo-ui/src/app/features/models/dummy/views/basic/FoolBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class FoolBasicView {

id: string;
selected: boolean;
canDelete: boolean;
code: string;
description: string;
longField: number;
booleanField: boolean;
doubleField: number;
decimalField: number;
dateField: string;
datetimeField: Date;

/* Specific Code Start */
/* Specific Code End */
}