// demo-ui/src/app/features/models/dummy/filters/FoolFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class FoolFilter {

code: string;
description: string;
booleanField: boolean;
dateFieldMinValue: string;
dateFieldMaxValue: string;
datetimeFieldMinValue: Date;
datetimeFieldMaxValue: Date;

/* Specific Code Start */
/* Specific Code End */
}