// demo-ui/src/app/features/models/dummy/sortings/FoolSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class FoolSorting {

codeOrderType: string;
descriptionOrderType: string;
longFieldOrderType: string;
booleanFieldOrderType: string;
doubleFieldOrderType: string;
decimalFieldOrderType: string;
dateFieldOrderType: string;
datetimeFieldOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}